+++ docs/changelog.md
@@ -71,6 +71,10 @@
 
 ## 更新说明详情
 
+1.12.65 fix: fix futures_foreign_commodity_realtime interface
+
+    1. 修复 futures_foreign_commodity_realtime 接口
+
 1.12.64 fix: fix stock_board_industry_index_ths interface
 
     1. 修复 stock_board_industry_index_ths 接口
@@ -3567,6 +3571,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.12.65 fix: fix futures_foreign_commodity_realtime interface
 
 1.12.64 fix: fix stock_board_industry_index_ths interface